Reports e2e coverage
