+++ .circleci/config.yml
@@ -46,6 +46,11 @@ commands:
           keys:
             - go-v1-{{ .Branch }}-{{ .Environment.CIRCLE_JOB }}
             - go-v1-master-{{ .Environment.CIRCLE_JOB }}
+  upload_codecov:
+    steps:
+      - run:
+          name: Uploading code coverage
+          command: bash <(curl -s https://codecov.io/bash) -f coverage.out
 jobs:
   codegen:
     docker:
@@ -96,9 +101,7 @@ jobs:
       - run: make test
       - save_vendor
       - save_go_cache
-      - run:
-          name: Uploading code coverage
-          command: bash <(curl -s https://codecov.io/bash) -f coverage.out
+      - upload_codecov
       - store_test_results:
           path: test-results
       - store_artifacts:
@@ -136,7 +139,7 @@ jobs:
       - restore_vendor
       - run: dep ensure
       - configure_git
-      - run: make cli
+      - run: make e2e-clis
       - run:
           name: Create namespace
           command: |
@@ -158,21 +161,20 @@ jobs:
           background: true
       - run:
           name: Start repo server
-          command: go run ./cmd/argocd-repo-server/main.go --loglevel debug --redis localhost:6379
+          command: ./dist/argocd-repo-server --loglevel debug --redis localhost:6379
           background: true
       - run:
           name: Start API server
-          command: go run ./cmd/argocd-server/main.go --loglevel debug --redis localhost:6379 --insecure --dex-server http://localhost:5556 --repo-server localhost:8081 --staticassets ../argo-cd-ui/dist/app
+          command: ./dist/argocd-server --loglevel debug --redis localhost:6379 --insecure --dex-server http://localhost:5556 --repo-server localhost:8081 --staticassets ../argo-cd-ui/dist/app
           background: true
       - run:
           name: Start Test Git
-          command: |
-            test/fixture/testrepos/start-git.sh
+          command: test/fixture/testrepos/start-git.sh
           background: true
       - run: until curl -v http://localhost:8080/healthz; do sleep 10; done
       - run:
           name: Start controller
-          command: go run ./cmd/argocd-application-controller/main.go --loglevel debug --redis localhost:6379 --repo-server localhost:8081 --kubeconfig ~/.kube/config
+          command: ./dist/argocd-application-controller --loglevel debug --redis localhost:6379 --repo-server localhost:8081 --kubeconfig ~/.kube/config
           background: true
       - run:
           command: PATH=dist:$PATH make test-e2e
@@ -182,6 +184,7 @@ jobs:
             ARGOCD_E2E_K3S: "true"
       - save_vendor
       - save_go_cache
+      - upload_codecov
       - store_test_results:
           path: test-results
       - store_artifacts:
