+++ test/e2e/cmd/argocd-server/root_test.go
@@ -1,0 +1,16 @@
+// +build e2efixtures
+
+package main
+
+import (
+	"testing"
+
+	"github.com/argoproj/argo-cd/cmd/argocd-server/commands"
+	"github.com/argoproj/argo-cd/test/e2e/cmd"
+)
+
+func TestArgoCDServer(t *testing.T) {
+	cmd.Wrap(func() error {
+		return commands.NewCommand().Execute()
+	})
+}
