+++ test/e2e/cmd/argocd/root_test.go
@@ -1,0 +1,16 @@
+// +build e2efixtures
+
+package main
+
+import (
+	"testing"
+
+	"github.com/argoproj/argo-cd/cmd/argocd/commands"
+	"github.com/argoproj/argo-cd/test/e2e/cmd"
+)
+
+func TestArgoCD(t *testing.T) {
+	cmd.Wrap(func() error {
+		return commands.NewCommand().Execute()
+	})
+}
